requirements.txt
django                  ==3.0.4
scrapy                  ==2.0.0
scrapyd                 ==1.2.1
python-scrapyd-api      ==2.1.2
sqlalchemy              ==1.3.15
django-tables2          ==2.3.1
openpyxl                ==3.0.3

pywin32
scrapy-user-agents
scrapy-rotating-proxies